+++ packages/jimtl_localazy/example/main.dart
@@ -1,7 +1,16 @@
+import 'package:jimtl/jimtl.dart';
+import 'package:jimtl_localazy/jimtl_localazy.dart';
+
 main() async {
-  //final manager = LocalazyCdnManager('_a860213072293210453319c546c5', 'errors.arb');
+  final manager = LocalazyCdnManager(
+    cdnId: 'CDI_ID_HERE',
+    cacheFolder: '.',
+    getFileName: (String locale, String flavor) {
+      return 'error.arb';
+    },
+  );
 
-  //print(await manager.download('en', 'flavor1'));
+  print(await manager.download('en', 'flavor1'));
 
-  //print(await manager.download('es_MX', IntlDelegate.defaultFlavorName));
+  print(await manager.download('es_MX', IntlDelegate.defaultFlavorName));
 }