+++ packages/flutter_jimtl/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.0.2
+
+* Add remote localization manager support for OTA ARB update
+
 ## 0.0.1
 
 * Initial release.